fix(quality): resolve SonarCloud new-code findings blocking the dev gate
The SonarCloud new-issue gate on dev failed with 56 findings, all
pre-existing code smells that count as new code in the dev->main diff.

- python:S9073 (48x): split composite `assert A and B` conjunctions in test
  modules into one assertion per condition, preserving assertion messages.
- python:S104 (1x): src/aptl/utils/pathsafe.py exceeded the 500-line file
  budget. Split into an underscore-sibling trio (matching the repo
  convention): _pathsafe_core (validation + walk primitives), _pathsafe_read
  (open/read/listdir), and pathsafe (the create-once write path, kept
  co-located with write_all so the test_pathsafe monkeypatch still resolves).
  The public import surface is unchanged and re-exported via __all__. Also
  drops the dead, never-referenced _open_leaf_create_exclusive_nofollow.
- python:S1128 (1x): the AptlRealization return annotation in
  _live_gate_probes was quoted, so it read as unused under
  `from __future__ import annotations`. Unquoted it so type-checkers and
  Sonar both see the use.
- python:S9084 (2x): removed the `import pytest as _pytest` aliases in
  test_runstore, importing pytest at module scope instead.
- python:S9077 (2x): replaced `assert False, msg` with `pytest.fail(msg)`.
- python:S9002 (2x): the two hypothesis @given fuzz tests use
  tmp_path_factory deliberately - a function-scoped tmp_path raises
  FailedHealthCheck and is shared across generated examples. Rather than
  suppress a false positive, switched to a tempfile.TemporaryDirectory
  context manager (the replacement hypothesis itself recommends), which
  mints a fresh store per example and drops the flagged fixture.

All affected tests pass (763 passed), including the pathsafe monkeypatch
tests and both fuzz tests under -m fuzz; the ruff C901 complexity gate and
all pathsafe consumers' imports verified.
