+++ tests/test_network_boundary_helper.py
@@ -14,7 +14,8 @@ HELPER = (
 @pytest.fixture(scope="module")
 def helper():
     spec = importlib.util.spec_from_file_location("network_boundary_helper", HELPER)
-    assert spec is not None and spec.loader is not None
+    assert spec is not None
+    assert spec.loader is not None
     module = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(module)
     return module
